refactor: Mongo 인덱스 생성 로직 이동
